♻️ [Refactoring]: #455 Phase 5 octal カテゴリを tests/octal.rs に分離（末尾 2 件編入）
8 進エスケープ greediness テスト 7 件に加え、末尾の別ブロックにあった
read_literal_string_decodes_octal_400_wraps_to_zero と
read_literal_string_decodes_octal_000_three_digit_zero の 2 件を
literal_string/tests/octal.rs に統合し、tests.rs ハブに `mod octal;` を追加。
